Move custom Input.delete() code to pre_delete in signals :arrow_right:
We need the delete Input signal to fire pre_delete in order to access the TaskIngredient info before it gets deleted. However, pre_delete fires before the DEFAULT delete call, not before we deleted TaskIngredient in custom code. Hence the move so we can put the cost update before we delete the TaskIngredient.