[BATCH SIZE WORKS!!] Create custom batch_size prev/new cost calculator
Since we assign cost predictably (first come, first serve) to children, when batch_size changes and we have to update costs, we now do it in the same order (first to last input). This is because there's the ing amount a child DESIRES to have, and the amount which the parent actually has remaining to contribute.
This is not an issue when updating the cost of something when the updated tasks batch size is unchanged. But we here had to write a custom function for batch_size changes.