Fix input delete (and, relatedly, task delete) by calling input delete @end of signal
Note that we're now returning an *optimistic* TaskIngredient list (gathered BEFORE running the async function), since we're not sure when the async function will finish and update the Task Ingredient. To do that we set an optional flag on a function that previously updated TaskIngredients, so we can just get the value we want to optimistically return.